requirements.txt
pandas 
numpy 
shutil 

jupyterlab
ipykernel


rstr 
facker

psycopg2 
sqlalchemy

pyaml 





